+++ src/main/java/com/irurueta/geometry/Polygon3D.java
@@ -77,7 +77,7 @@ public class Polygon3D implements Serializable {
     /**
      * Constructor.
      *
-     * @param vertices List of vertices forming this polygon..
+     * @param vertices List of vertices forming this polygon.
      * @throws NotEnoughVerticesException Raised if list does not contain enough
      *                                    vertices.
      * @see #MIN_VERTICES
@@ -322,9 +322,9 @@ public class Polygon3D implements Serializable {
         final int total = mVertices.size();
 
         for (final Point3D point : mVertices) {
-            inhomX += point.getInhomX() / (double) total;
-            inhomY += point.getInhomY() / (double) total;
-            inhomZ += point.getInhomZ() / (double) total;
+            inhomX += point.getInhomX() / total;
+            inhomY += point.getInhomY() / total;
+            inhomZ += point.getInhomZ() / total;
         }
         result.setInhomogeneousCoordinates(inhomX, inhomY, inhomZ);
     }
@@ -426,7 +426,7 @@ public class Polygon3D implements Serializable {
             // to increase accuracy
             line.normalize();
 
-            // find closest point to line
+            // find the closest point to line
             line.closestPoint(point, pointInLine);
             // to increase accuracy
             pointInLine.normalize();
@@ -460,7 +460,7 @@ public class Polygon3D implements Serializable {
         // to increase accuracy
         line.normalize();
 
-        // find closest point to line
+        // find the closest point to line
         line.closestPoint(point, pointInLine);
         // to increase accuracy
         pointInLine.normalize();
@@ -479,7 +479,7 @@ public class Polygon3D implements Serializable {
 
         if (!found) {
             // no closest point was found on a segment belonging to polygon
-            // boundary so we search for the closest vertex
+            // boundary, so we search for the closest vertex
             iterator = mVertices.iterator();
             while (iterator.hasNext()) {
                 // a better vertex has been found
@@ -553,7 +553,7 @@ public class Polygon3D implements Serializable {
             // to increase accuracy
             line.normalize();
 
-            // find closest point to line
+            // find the closest point to line
             line.closestPoint(point, pointInLine);
             // to increase accuracy
             pointInLine.normalize();
@@ -589,7 +589,7 @@ public class Polygon3D implements Serializable {
         // to increase accuracy
         line.normalize();
 
-        // find closest point to line
+        // find the closest point to line
         line.closestPoint(point, pointInLine);
         // to increase accuracy
         pointInLine.normalize();
@@ -609,7 +609,7 @@ public class Polygon3D implements Serializable {
 
         if (!found) {
             // no closest point was found on a segment belonging to polygon
-            // boundary so we search for the closest vertex
+            // boundary, so we search for the closest vertex
             iterator = mVertices.iterator();
             while (iterator.hasNext()) {
                 curPoint = iterator.next();
